feat: migrate domain module to KMP
