+++ domain/src/commonMain/kotlin/com/mindeck/domain/exception/DomainError.kt
