+++ domain/src/commonMain/kotlin/com/mindeck/domain/models/CardState.kt
