+++ domain/src/commonMain/kotlin/com/mindeck/domain/usecases/card/query/GetCardByIdUseCase.kt
@@ -1,0 +1,9 @@
+package com.mindeck.domain.usecases.card.query
+
+import com.mindeck.domain.models.Card
+import com.mindeck.domain.repository.CardRepository
+import kotlinx.coroutines.flow.Flow
+
+class GetCardByIdUseCase(private val cardRepository: CardRepository) {
+    operator fun invoke(cardId: Int): Flow<Card?> = cardRepository.getCardById(cardId = cardId)
+}
