+++ domain/src/commonMain/kotlin/com/mindeck/domain/usecases/deck/command/CreateDeckUseCase.kt
@@ -1,0 +1,10 @@
+package com.mindeck.domain.usecases.deck.command
+
+import com.mindeck.domain.models.Deck
+import com.mindeck.domain.repository.DeckRepository
+
+class CreateDeckUseCase(private val deckRepository: DeckRepository) {
+    suspend operator fun invoke(deck: Deck): Int {
+        return deckRepository.insertDeck(deck = deck)
+    }
+}
