+++ domain/src/commonMain/kotlin/com/mindeck/domain/usecases/deck/command/DeleteDeckUseCase.kt
@@ -1,0 +1,9 @@
+package com.mindeck.domain.usecases.deck.command
+
+import com.mindeck.domain.repository.DeckRepository
+
+class DeleteDeckUseCase(private val deckRepository: DeckRepository) {
+    suspend operator fun invoke(deckId: Int) {
+        deckRepository.deleteDeck(deckId = deckId)
+    }
+}
