+++ domain/src/commonMain/kotlin/com/mindeck/domain/usecases/deck/query/GetAllDecksUseCase.kt
@@ -1,0 +1,11 @@
+package com.mindeck.domain.usecases.deck.query
+
+import com.mindeck.domain.models.Deck
+import com.mindeck.domain.repository.DeckRepository
+import kotlinx.coroutines.flow.Flow
+
+class GetAllDecksUseCase(private val deckRepository: DeckRepository) {
+    operator fun invoke(): Flow<List<Deck>> {
+        return deckRepository.getAllDecks()
+    }
+}
