--- domain/src/main/java/com/mindeck/domain/usecases/card/query/GetCardByIdUseCase.kt
@@ -1,10 +1,0 @@
-package com.mindeck.domain.usecases.card.query
-
-import com.mindeck.domain.models.Card
-import com.mindeck.domain.repository.CardRepository
-import kotlinx.coroutines.flow.Flow
-import javax.inject.Inject
-
-class GetCardByIdUseCase @Inject constructor(private val cardRepository: CardRepository) {
-    operator fun invoke(cardId: Int): Flow<Card?> = cardRepository.getCardById(cardId = cardId)
-}
