--- domain/src/main/java/com/mindeck/domain/usecases/deck/command/CreateDeckUseCase.kt
@@ -1,11 +1,0 @@
-package com.mindeck.domain.usecases.deck.command
-
-import com.mindeck.domain.models.Deck
-import com.mindeck.domain.repository.DeckRepository
-import javax.inject.Inject
-
-class CreateDeckUseCase @Inject constructor(private val deckRepository: DeckRepository) {
-    suspend operator fun invoke(deck: Deck): Int {
-        return deckRepository.insertDeck(deck = deck)
-    }
-}
